Move most of the network code to agama-lib
