+++ rust/agama-lib/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod error;
 pub mod install_settings;
 pub mod manager;
+pub mod network;
 pub mod profile;
 pub mod settings;
 pub mod software;
