+++ rust/agama-lib/src/network.rs
@@ -1,0 +1,5 @@
+pub mod dbus;
+pub mod state;
+
+pub use dbus::NetworkService;
+pub use state::{NetworkState, NetworkStateError};
