+++ rust/agama-lib/src/network/state.rs
